requirements.txt
transformers
torchvision
torch
Pillow
accelerate
sentencepiece
protobuf
huggingface_hub
numpy
matplotlib
timm
einops